Mock response test cases
